requirements.txt
numpy>=1.24.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
faiss-cpu>=1.7.4 ; platform_system != 'Darwin'  # macOS may struggle in CI, optional 